JP-2015: MIRI & NIRSpec 3D drizzling C extension (#6297)
* MIRI 3D drizzling c extension

* Take out normalizing area by pixel area

* updates for NIRSpec blotting

* start at getting drizzle for NIRSPEC

* update for NIRSPEC dwave

* update

* test a few things with NIRSpec

* added NIRSpec driz

* removed python 3d drizzing test code

* changes from review

* removed all code related to miri_psf weighting

* clean up doc

* flake fix

* Changes from MC review

* no change

* trying to fix build errors
