+++ jwst/cube_build/cube_build_io_util.py
@@ -28,7 +28,7 @@ def read_cubepars(par_filename,
     instrument : str
         Either MIRI or NIRSPEC
     weighting : str
-        Type of weighting, msm, emem or miripsf
+        Type of weighting, msm, emem or drizzle
     all_channel : list
         all the channels contained in input data
     all_subchannel : list
@@ -99,7 +99,7 @@ def read_cubepars(par_filename,
                                                     table_sroi, table_wroi,
                                                     table_scalerad)
 
-            #  read in wavelength table for modified shepard method 1/r weighting
+            #  read in wavelength table for multi-band data
             if weighting == 'msm':
                 for tabdata in ptab.ifucubepars_multichannel_msm_wavetable:
                     table_wave = tabdata['WAVELENGTH']
@@ -111,7 +111,7 @@ def read_cubepars(par_filename,
                                                          table_wroi, table_power,
                                                          table_softrad)
             #  read in wavelength table for modified shepard method 1/r weighting
-            elif weighting == 'emsm' or weighting == 'miripsf':
+            elif weighting == 'emsm':
                 for tabdata in ptab.ifucubepars_multichannel_emsm_wavetable:
                     table_wave = tabdata['WAVELENGTH']
                     table_sroi = tabdata['ROISPATIAL']
@@ -119,6 +119,10 @@ def read_cubepars(par_filename,
                     table_scalerad = tabdata['SCALERAD']
                     instrument_info.SetMultiChannelEMSMTable(table_wave, table_sroi,
                                                              table_wroi, table_scalerad)
+            elif weighting == 'drizzle':
+                for tabdata in ptab.ifucubepars_multichannel_driz_wavetable:
+                    table_wave = tabdata['WAVELENGTH']
+                    instrument_info.SetMultiChannelDrizTable(table_wave)
 
     # Read in NIRSPEC Values
     elif instrument == 'NIRSPEC':
@@ -226,97 +230,3 @@ def read_cubepars(par_filename,
                     table_scalerad = tabdata['SCALERAD']
                     instrument_info.SetHighEMSMTable(table_wave, table_sroi,
                                                      table_wroi, table_scalerad)
-# _____________________________________________________________________________
-
-
-def read_resolution_file(resol_filename,
-                         all_channel,
-                         all_subchannel,
-                         instrument_info):
-    """ Read in the MIRI Resolution reference file
-
-    If this is MIRI data and the weighting is miripsf then read in the
-    MIRI resolition file. Read weighting parameters based on which channel
-    and subchannel we are working with and fill in the appropriated values
-    into instrument_info.
-
-    Parameters
-    ----------
-    resol_filename : str
-        MIRI resolution reference table
-    all_channel : list
-        all the channels contained in input data
-    all_subchannel : list
-        all subchannels contained in input data
-    instrument_info : dictionary
-        holds the  MIRI psf weighting parameters
-
-    Returns
-    -------
-    Resolution parameters are filled in in the instrumeent_info
-    dectionary
-
-    """
-
-    ptab = datamodels.MiriResolutionModel(resol_filename)
-    table_alpha_cutoff = ptab.psf_fwhm_alpha_table['A_CUTOFF']
-    table_alpha_a_short = ptab.psf_fwhm_alpha_table['A_A_SHORT']
-    table_alpha_b_short = ptab.psf_fwhm_alpha_table['A_B_SHORT']
-    table_alpha_a_long = ptab.psf_fwhm_alpha_table['A_A_LONG']
-    table_alpha_b_long = ptab.psf_fwhm_alpha_table['A_B_LONG']
-
-    table_beta_cutoff = ptab.psf_fwhm_beta_table['B_CUTOFF']
-    table_beta_a_short = ptab.psf_fwhm_beta_table['B_A_SHORT']
-    table_beta_b_short = ptab.psf_fwhm_beta_table['B_B_SHORT']
-    table_beta_a_long = ptab.psf_fwhm_beta_table['B_A_LONG']
-    table_beta_b_long = ptab.psf_fwhm_beta_table['B_B_LONG']
-
-    instrument_info.Set_psf_alpha_parameters(table_alpha_cutoff,
-                                             table_alpha_a_short,
-                                             table_alpha_b_short,
-                                             table_alpha_a_long,
-                                             table_alpha_b_long)
-
-    instrument_info.Set_psf_beta_parameters(table_beta_cutoff,
-                                            table_beta_a_short,
-                                            table_beta_b_short,
-                                            table_beta_a_long,
-                                            table_beta_b_long)
-
-    number_bands = len(all_channel)
-    # pull out the channels and subcahnnels that the cube is make from
-    for i in range(number_bands):
-        this_channel = all_channel[i]
-        this_sub = all_subchannel[i]
-        compare_band = this_channel + this_sub
-        for tabdata in ptab.resolving_power_table:
-            table_sub_band = tabdata['SUB_BAND'].lower()
-            table_wave_center = tabdata['R_CENTRE']
-            table_res_a_low = tabdata['R_A_LOW']
-            table_res_b_low = tabdata['R_B_LOW']
-            table_res_c_low = tabdata['R_C_LOW']
-            table_res_a_high = tabdata['R_A_HIGH']
-            table_res_b_high = tabdata['R_B_HIGH']
-            table_res_c_high = tabdata['R_C_HIGH']
-            table_res_a_ave = tabdata['R_A_AVG']
-            table_res_b_ave = tabdata['R_B_AVG']
-            table_res_c_ave = tabdata['R_C_AVG']
-            # match on this_channel and this_sub
-            if compare_band == table_sub_band:
-                instrument_info.Set_RP_Wave_Cutoff(table_wave_center,
-                                                   this_channel, this_sub)
-
-                instrument_info.Set_RP_low(table_res_a_low,
-                                           table_res_b_low,
-                                           table_res_c_low,
-                                           this_channel, this_sub)
-
-                instrument_info.Set_RP_high(table_res_a_high,
-                                            table_res_b_high,
-                                            table_res_c_high,
-                                            this_channel, this_sub)
-
-                instrument_info.Set_RP_ave(table_res_a_ave,
-                                           table_res_b_ave,
-                                           table_res_c_ave,
-                                           this_channel, this_sub)
