+++ jwst/cube_build/src/blot_median.c
@@ -57,7 +57,7 @@ blot_weight : numpy.ndarray
 #include <numpy/arrayobject.h>
 #include <numpy/npy_math.h>
 
-#define PY_ARRAY_UNIQUE_SYMBOL _jwst_cube_match_sky_numpy_api    //WHAT IS THIS AND WHERE IS IT USED???
+#define PY_ARRAY_UNIQUE_SYMBOL _jwst_cube_blot_numpy_api
 #define NPY_NO_DEPRECATED_API NPY_1_7_API_VERSION
 
 //_______________________________________________________________________
